+++ lib/check_license.dart
@@ -1,0 +1,5 @@
+export 'package:license_checker/src/config.dart';
+export 'package:license_checker/src/package_checker.dart';
+export 'package:license_checker/src/dependency_checker.dart'
+    show DependencyChecker;
+export 'package:license_checker/src/check_license.dart';
