+++ lib/generate_disclaimer.dart
@@ -1,0 +1,3 @@
+export 'package:license_checker/src/config.dart';
+export 'package:license_checker/src/package_checker.dart';
+export 'package:license_checker/src/generate_disclaimer.dart';
