+++ test/lib/src/generate_disclaimer_test.dart
@@ -5,17 +5,10 @@ import 'package:test/test.dart';
 
 import 'package:mockito/mockito.dart';
 
-import 'package:license_checker/src/checker.dart';
+import 'package:license_checker/src/generate_disclaimer.dart';
 import 'package:license_checker/src/config.dart';
 import 'package:license_checker/src/package_checker.dart';
 import 'package:license_checker/src/dependency_checker.dart';
-
-// class MockedPackage extends Mock implements Package {
-//   @override
-//   final Package package;
-
-//   MockedDependencyChecker(this.package);
-// }
 
 class MockedDependencyChecker extends Mock implements DependencyChecker {
   @override
@@ -64,153 +57,6 @@ class MockedPackageChecker extends Mock implements PackageChecker {
 }
 
 void main() {
-  group('Check License', () {
-    String licenseDisplay({
-      required String packageName,
-      required LicenseStatus licenseStatus,
-      required String licenseName,
-    }) {
-      return '$packageName $licenseName ${licenseStatus.toString()}';
-    }
-
-    test('should check the license and return the display for one package',
-        () async {
-      DependencyChecker package =
-          MockedDependencyChecker('Dodgers', LicenseStatus.approved);
-
-      expect(
-        await checkPackageLicense(
-          package: package,
-          licenseDisplay: licenseDisplay,
-        ),
-        'Dodgers swag LicenseStatus.approved',
-      );
-    });
-
-    test('should check the license of all dependencies of a package', () async {
-      PackageChecker packageConfig = MockedPackageChecker([
-        MockedDependencyChecker('Dodgers', LicenseStatus.approved),
-        MockedDependencyChecker('Giants', LicenseStatus.rejected),
-      ]);
-      List<LicenseDisplayWithPriority<String>> result =
-          await checkAllPackageLicenses(
-        packageConfig: packageConfig,
-        showDirectDepsOnly: false,
-        filterApproved: false,
-        licenseDisplay: licenseDisplay,
-      );
-
-      expect(result.map((e) => e.display.toString()), [
-        'Dodgers swag LicenseStatus.approved',
-        'Giants swag LicenseStatus.rejected',
-      ]);
-
-      expect(result.map((e) => e.status), [
-        LicenseStatus.approved,
-        LicenseStatus.rejected,
-      ]);
-    });
-
-    test('should check the license of direct dependencies of a package',
-        () async {
-      PackageChecker packageConfig = MockedPackageChecker([
-        MockedDependencyChecker('Dodgers', LicenseStatus.approved),
-        MockedDependencyChecker('Giants', LicenseStatus.rejected),
-        MockedDependencyChecker('Kings', LicenseStatus.approved),
-      ]);
-      List<LicenseDisplayWithPriority<String>> result =
-          await checkAllPackageLicenses(
-        packageConfig: packageConfig,
-        showDirectDepsOnly: true,
-        filterApproved: false,
-        licenseDisplay: licenseDisplay,
-      );
-
-      expect(result.map((e) => e.display.toString()), [
-        'Dodgers swag LicenseStatus.approved',
-        'Giants swag LicenseStatus.rejected',
-      ]);
-
-      expect(result.map((e) => e.status), [
-        LicenseStatus.approved,
-        LicenseStatus.rejected,
-      ]);
-    });
-
-    test(
-        'should check the license of all dependencies of a package and filter out approved and permitted license',
-        () async {
-      PackageChecker packageConfig = MockedPackageChecker([
-        MockedDependencyChecker('Dodgers', LicenseStatus.approved),
-        MockedDependencyChecker('Giants', LicenseStatus.rejected),
-        MockedDependencyChecker('Angles', LicenseStatus.unknown),
-        MockedDependencyChecker('As', LicenseStatus.permitted),
-        MockedDependencyChecker('Padres', LicenseStatus.needsApproval),
-        MockedDependencyChecker('Rockies', LicenseStatus.noLicense),
-      ]);
-      List<LicenseDisplayWithPriority<String>> result =
-          await checkAllPackageLicenses(
-        packageConfig: packageConfig,
-        showDirectDepsOnly: false,
-        filterApproved: true,
-        licenseDisplay: licenseDisplay,
-      );
-
-      expect(result.map((e) => e.display.toString()), [
-        'Giants swag LicenseStatus.rejected',
-        'Angles swag LicenseStatus.unknown',
-        'Padres swag LicenseStatus.needsApproval',
-        'Rockies swag LicenseStatus.noLicense',
-      ]);
-
-      expect(result.map((e) => e.status), [
-        LicenseStatus.rejected,
-        LicenseStatus.unknown,
-        LicenseStatus.needsApproval,
-        LicenseStatus.noLicense,
-      ]);
-    });
-
-    test(
-        'should check the license of all dependencies and sort them correctly by status priority',
-        () async {
-      PackageChecker packageConfig = MockedPackageChecker([
-        MockedDependencyChecker('Dodgers', LicenseStatus.approved),
-        MockedDependencyChecker('Giants', LicenseStatus.rejected),
-        MockedDependencyChecker('Angles', LicenseStatus.unknown),
-        MockedDependencyChecker('As', LicenseStatus.permitted),
-        MockedDependencyChecker('Padres', LicenseStatus.needsApproval),
-        MockedDependencyChecker('Rockies', LicenseStatus.noLicense),
-      ]);
-      List<LicenseDisplayWithPriority<String>> result =
-          await checkAllPackageLicenses(
-        packageConfig: packageConfig,
-        showDirectDepsOnly: false,
-        filterApproved: false,
-        licenseDisplay: licenseDisplay,
-        sort: true,
-      );
-
-      expect(result.map((e) => e.display.toString()), [
-        'Dodgers swag LicenseStatus.approved',
-        'As swag LicenseStatus.permitted',
-        'Angles swag LicenseStatus.unknown',
-        'Rockies swag LicenseStatus.noLicense',
-        'Padres swag LicenseStatus.needsApproval',
-        'Giants swag LicenseStatus.rejected',
-      ]);
-
-      expect(result.map((e) => e.status), [
-        LicenseStatus.approved,
-        LicenseStatus.permitted,
-        LicenseStatus.unknown,
-        LicenseStatus.noLicense,
-        LicenseStatus.needsApproval,
-        LicenseStatus.rejected,
-      ]);
-    });
-  });
-
   group('Generate Disclaimer', () {
     String disclaimerCLIDisplay({
       required String packageName,
@@ -315,7 +161,8 @@ void main() {
       expect(
         result.file[0],
         equals(
-            'file: Dodgers swag 1958 Los Angeles Chavez Ravine /dodger/stadium'),
+          'file: Dodgers swag 1958 Los Angeles Chavez Ravine /dodger/stadium',
+        ),
       );
       expect(
         result.file[1],
